+++ lib/impl/_impl.dart
@@ -1,7 +1,7 @@
-export 'cache.dart';
 export 'clipboard.dart';
 export 'ev_center.dart';
 export 'external_link.dart';
+export 'file_cache.dart';
 export 'image_op.dart';
 export 'kv_storage.dart';
 export 'meta.dart';